fix: 판매글 리스트 최신순 정렬로 수정
판매글 수정시, 글을 작성한 사람만 수정 가능하도록 변경

프로필 사진 삭제시, 유효성 검사 추가

+++ src/main/java/com/sparta/market/domain/user/service/UserService.java
@@ -149,9 +149,16 @@ public class UserService {
         User findUser = userRepository.findById(user.getId()).orElseThrow(()->
                 new CustomException(ErrorCode.NOT_EXIST_USER)
         );
+        UserProfile userProfile = userProfileRepository.findByUserId(user.getId()).orElseThrow(() ->
+                new CustomException(ErrorCode.NOT_EXIST_PROFILE)
+        );
         UserProfile profile = userProfileRepository.findById(profileId).orElseThrow(()->
                 new CustomException(ErrorCode.NOT_EXIST_PROFILE)
         );
+        if (profile.getUser().getId() != userProfile.getUser().getId()) {
+            throw new CustomException(ErrorCode.NOT_YOUR_IMG);
+        }
+
 
         s3UploadService.deleteFile(profile.getS3name());
         userProfileRepository.delete(profile);
